requirements.txt
PyPDF2
openai
PyYAML